chore(api/user): segregate functions in files
